requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
torch>=1.12.0
transformers>=4.20.0
datasets>=2.0.0

# Machine Learning Libraries
lightgbm>=3.3.0
xgboost>=1.6.0
catboost>=1.0.0

# Data Processing
openpyxl>=3.0.0
xlrd>=2.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
joblib>=1.1.0
loguru>=0.6.0

# Development and Testing (optional)
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
pre-commit>=2.15.0

# Documentation (optional)
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Jupyter support (optional)
jupyter>=1.0.0
ipywidgets>=7.6.0